What's on the menu?
Added a menu page, with options to start the game, quit it and to read
the instructions of the game.  Also added a back button for going back
to the main menu and an alert box to warn the user about quitting the
game, which will happen when the back button is pressed.