Animation? Complete(with a lot of other things)!
Added all the animations to required functions and they all work just
fine! Changed Opacity = 0 in the BigWin function to Visibility =
Visibility.Collapsed as nullifying opacity became a bug due to the
animation of the top three grids, which required to change the opacity
to zero and back to full each time they are triggered. Hence, collapsing
visibility seemed a better idea. Apart from this, added proper splash
screen to the phone version(Lumia 520 as of now) and all the proper
offline square icons(device local ones, that is). Yet, there are three
major bugs because of the three grids of the first row appearing after
each selection of the grids of the second row. Gotta fix this soon.