Eureka!
Finally fixed the opacity bug that's been bugging me for a while. Also
fixed the disappearance of the won symbol in block 1,2. The latter was
caused by a small error in the arrangement of animations in Blend, which
slipped under my nose.  Fixed the opacity bug by changing the code
inside the "Won" method which sets the opacity of the elements in the
won grid to '0', to collapsing it as soon as the block is won. One more
major bug remains: It's taking forever to restart the game, after
clicking on the "Play again" button.